+++ src/biz/bokhorst/xprivacy/XPrivacy.java
@@ -1,6 +1,5 @@
 package biz.bokhorst.xprivacy;
 
-import java.io.File;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
@@ -15,7 +14,6 @@ import android.app.AndroidAppHelper;
 import android.content.Context;
 import android.os.Binder;
 import android.os.Build;
-import android.os.Environment;
 import android.os.Process;
 import android.util.Log;
 
@@ -29,7 +27,6 @@ import static de.robv.android.xposed.XposedHelpers.findClass;
 
 @SuppressLint("DefaultLocale")
 public class XPrivacy implements IXposedHookLoadPackage, IXposedHookZygoteInit {
-	private static boolean mLBE = false;
 	private static String mSecret = null;
 	private static boolean mAccountManagerHooked = false;
 	private static boolean mActivityManagerHooked = false;
@@ -48,15 +45,6 @@ public class XPrivacy implements IXposedHookLoadPackage, IXposedHookZygoteInit {
 	@SuppressLint("InlinedApi")
 	public void initZygote(StartupParam startupParam) throws Throwable {
 		Util.log(null, Log.INFO, String.format("Load %s", startupParam.modulePath));
-
-		// Check for LBE Security Guard (incompatible)
-		File apps = new File(Environment.getDataDirectory() + File.separator + "app");
-		for (File file : apps.listFiles())
-			if (file.getName().startsWith("com.lbe.security")) {
-				mLBE = true;
-				Util.log(null, Log.WARN, "LBE detected, not hooking");
-				return;
-			}
 
 		// Generate secret
 		mSecret = Long.toHexString(new Random().nextLong());
@@ -158,9 +146,6 @@ public class XPrivacy implements IXposedHookLoadPackage, IXposedHookZygoteInit {
 	}
 
 	public void handleLoadPackage(final LoadPackageParam lpparam) throws Throwable {
-		if (mLBE)
-			return;
-
 		// Log load
 		Util.log(null, Log.INFO, String.format("Load package=%s uid=%d", lpparam.packageName, Process.myUid()));
 
@@ -203,9 +188,6 @@ public class XPrivacy implements IXposedHookLoadPackage, IXposedHookZygoteInit {
 	}
 
 	public static void handleGetSystemService(XHook hook, String name, Object instance) {
-		if (mLBE)
-			return;
-
 		Util.log(hook, Log.INFO,
 				"getSystemService " + name + "=" + instance.getClass().getName() + " uid=" + Binder.getCallingUid());