+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+* Compatibility with LBE Security Master ([issue](https://github.com/M66B/XPrivacy/issues/1167))
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.99.23 BETA**